Extract AC05 health schema into dedicated microcrate
