+++ Cargo.toml
@@ -31,6 +31,7 @@ members = [
   "crates/bitnet-intel-gpu-id",  # SRP: Intel GPU/Arc identification heuristics
   "crates/bitnet-logits",        # SRP: pure logits transform functions
   "crates/bitnet-logits-filters", # SRP: reusable logits/probability filtering transforms
+  "crates/bitnet-ac05-types",    # SRP: shared AC05 health response schema types
   "crates/bitnet-generation",    # SRP: decode-loop stopping logic & generation events
   "crates/bitnet-generation-stop-core", # SRP: reusable stop criteria/check core
   "crates/bitnet-engine-core",   # SRP: orchestration contracts (traits, session types)
@@ -131,6 +132,7 @@ default-members = [
   "crates/bitnet-device-probe",
   "crates/bitnet-logits",
   "crates/bitnet-logits-filters",
+  "crates/bitnet-ac05-types",
   "crates/bitnet-generation",
   "crates/bitnet-generation-stop-core",
   "crates/bitnet-engine-core",
